Backend/Fast-api/requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
python-dotenv
nest_asyncio
gdown
opencv-python-headless
pandas
numpy
requests